em-parity: K=2 — promote per-image stats to float64 (closes class_acc 0.54 → 0.94)
In the K-class CC mode (relion_firstiter_score_mode='normalized_cc'),
recovar's K-class joint argmax over best_log_score_per_image collapsed
because the float32 stats buffer rounded the per-class CC delta to 0.

The arithmetic: log_evidence_per_image = log_Z + log_score_offset, where
log_score_offset = -0.5 * Σ |Fimg|²/σ² · hw is the image-power normalizer
(magnitude ~5e6 for a 128² K=2 fixture). The CC-mode score is bounded in
[-1, 1] and the per-class delta is ~0.06 in image-power units after the
1/Xi2 rescaling. float32 precision at ±5e6 is ~0.6, so adding the small
delta to the large offset rounded the entire per-class difference to 0
for >75% of particles. Per-class best_log_score values matched bit for bit.

K-class _assemble_result then did argmax(best_score) over identical values:
the tie-break became deterministic and assigned 4981/5000 particles to
class 0, dropping K=2 fast test class_assignment_accuracy to 0.5398.

Fix: allocate log_evidence_per_image and best_log_score_per_image as
float64 in run_em, and accumulate the (best_score + log_score_offset)
sum in float64 before the cast. JAX's x64 mode is already enabled
(recovar/jax_config.py:13), so the make_relion_stats helper preserves
float64 through the K-class assembly path. max_posterior_per_image stays
in float32 — its magnitude is in [0, 1] with no large-offset addition.

Validation:
  K=2 fast test (test_em_parity_fast_kclass_replay):
    class_assignment_accuracy 0.5398 → 0.9364 (+39pp; threshold 0.95)
    mean_corr unchanged at 0.9843
    pmax_abs_mean unchanged at 0.0
    diff (c0_best_log_score - c1_best_log_score) std: 0 → 0.015
    recovar c0/c1 split 4981/19 → 2708/2292 (matches RELION 2680/2320)

  K=1 fast test (test_em_parity_fast_k1_replay):
    half1 corr 0.999951 → 0.999955 (unchanged)
    pmax_abs_diff unchanged

The remaining 6.4% K=2 misassignment gap is a true algorithmic divergence
(recovar's Hermitian doubling on suma2 vs RELION's unit-weighted sum at
ml_optimiser.cpp:8769; recovar's 1/sqrt(Xi2_hw)·1/sqrt(suma2_hw) vs
RELION's 1/sqrt(Xi2)·1/sqrt(suma2)) — a class-symmetric scale factor for
most poses, but tiny per-pose differences in the Hermitian-weighted vs
unweighted Re(<Fimg, Frefctf>) can flip argmax for the ~6% of particles
whose c0/c1 CC delta is below the algorithmic noise floor.

+++ recovar/em/dense_single_volume/em_engine.py
@@ -893,8 +893,13 @@ def run_em(
     max_posterior_per_image = None
     rotation_posterior_sums = None
     if return_stats:
-        log_evidence_per_image = np.empty(n_images, dtype=np.float32)
-        best_log_score_per_image = np.empty(n_images, dtype=np.float32)
+        # K-class joint argmax over best_log_score relies on the per-class
+        # delta surviving the addition of the (large) image-power log_score_offset.
+        # Float32 rounding at ±5e6 magnitudes quantizes those deltas to 0 and
+        # collapses the CC-mode K=2 class assignment, so keep the per-image
+        # log/score buffers in float64.
+        log_evidence_per_image = np.empty(n_images, dtype=np.float64)
+        best_log_score_per_image = np.empty(n_images, dtype=np.float64)
         max_posterior_per_image = np.empty(n_images, dtype=np.float32)
         rotation_posterior_sums = np.zeros(n_rot, dtype=np.float64)
 
@@ -1682,17 +1687,22 @@ def run_em(
                 pmax = jnp.ones_like(best_score)
             else:
                 pmax = jnp.exp(best_score - log_Z)
+            # Cast the per-image stats sums into the destination buffer dtype
+            # (float64 if the buffer was allocated as float64) so the small
+            # CC-mode delta between best_score values survives the addition of
+            # the much-larger image_power-scaled log_score_offset. Float32
+            # rounding here would quantize per-class K=2 differences to 0.
             log_evidence_per_image[start_idx:end_idx] = np.asarray(
-                log_Z_actual + log_score_offset,
-                dtype=np.float32,
+                (log_Z_actual.astype(jnp.float64) + log_score_offset.astype(jnp.float64)),
+                dtype=log_evidence_per_image.dtype,
             )
             best_log_score_per_image[start_idx:end_idx] = np.asarray(
-                best_score_actual + log_score_offset,
-                dtype=np.float32,
+                (best_score_actual.astype(jnp.float64) + log_score_offset.astype(jnp.float64)),
+                dtype=best_log_score_per_image.dtype,
             )
             max_posterior_per_image[start_idx:end_idx] = np.asarray(
                 pmax[:actual_batch_size],
-                dtype=np.float32,
+                dtype=max_posterior_per_image.dtype,
             )
             if sync_timers:
                 _block_until_ready(log_Z, best_score, pmax)
